+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/content/LinkAttachmentContent.kt
@@ -85,13 +85,50 @@ import io.getstream.chat.android.uiutils.extension.hasLink
  */
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
+@Deprecated(
+    message = "Use the version with the `onItemClick` parameter that accepts a LinkAttachmentClickData.",
+    replaceWith = ReplaceWith(
+        expression = "LinkAttachmentContent(attachmentState, linkDescriptionMaxLines, modifier, onItemClick)",
+    ),
+)
 public fun LinkAttachmentContent(
     attachmentState: AttachmentState,
     linkDescriptionMaxLines: Int,
     modifier: Modifier = Modifier,
-    onItemClick: (context: Context, url: String) -> Unit = ::onLinkAttachmentContentClick,
+    onItemClick: (context: Context, url: String) -> Unit,
 ) {
-    val (message, isMine, onLongItemClick) = attachmentState
+    LinkAttachmentContent(
+        state = attachmentState,
+        linkDescriptionMaxLines = linkDescriptionMaxLines,
+        modifier = modifier,
+        onItemClick = { onItemClick(it.context, it.url) },
+    )
+}
+
+/**
+ * Builds a link attachment message, which shows the link image preview, the title of the link
+ * as well as its description.
+ *
+ * When clicking it, we open the preview link.
+ *
+ * @param state - The state of the attachment, holding the root modifier, the message
+ * and the onLongItemClick handler.
+ * @param linkDescriptionMaxLines - The limit of how many lines we show for the link description.
+ * @param modifier Modifier for styling.
+ * @param onItemClick Lambda called when an item gets clicked.
+ */
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+@Suppress("LongMethod")
+public fun LinkAttachmentContent(
+    state: AttachmentState,
+    linkDescriptionMaxLines: Int,
+    modifier: Modifier = Modifier,
+    onItemClick: (LinkAttachmentClickData) -> Unit = {
+        onLinkAttachmentContentClick(it.context, it.url)
+    },
+) {
+    val (message, isMine, onLongItemClick) = state
 
     val context = LocalContext.current
     val attachment = message.attachments.firstOrNull { it.hasLink() && !it.isGiphy() }
@@ -119,7 +156,14 @@ public fun LinkAttachmentContent(
                 onClick = {
                     try {
                         if (urlWithScheme != null) {
-                            onItemClick(context, urlWithScheme)
+                            onItemClick(
+                                LinkAttachmentClickData(
+                                    context = context,
+                                    url = urlWithScheme,
+                                    attachment = attachment,
+                                    message = message,
+                                ),
+                            )
                         } else {
                             Toast
                                 .makeText(context, errorMessage, Toast.LENGTH_LONG)
@@ -291,13 +335,28 @@ internal fun LinkAttachmentContent() {
             authorName = "Author",
         )
         LinkAttachmentContent(
-            attachmentState = AttachmentState(
+            state = AttachmentState(
                 message = Message(attachments = listOf(attachment)),
             ),
             linkDescriptionMaxLines = 5,
         )
     }
 }
+
+/**
+ * Data class that holds information about a link attachment click event.
+ *
+ * @param context The context in which the click event occurred.
+ * @param url The URL of the link attachment that was clicked.
+ * @param message The message containing the link attachment.
+ */
+@ConsistentCopyVisibility
+public data class LinkAttachmentClickData internal constructor(
+    val context: Context,
+    val url: String,
+    val attachment: Attachment,
+    val message: Message,
+)
 
 @Suppress("MagicNumber")
 private val LongDescription = (0..50).joinToString { "Lorem ipsum dolor sit amet" }